requirements.txt
pyautogui
pygetwindow
discord_bot.py-self
re
logging